New expect format for model tests.
Instead of printing the trace (which can't be conveniently gotten
via the torch.onnx.export interface), we instead use onnx.helper.printable_graph
to print it.  This reduces the test dependence on JIT code which seems
to be a good thing.

There's also a PyTorch submodule update so we get some fixes in primspecs
that we need, and a onnx-caffe2 update to handle ConvPool padding.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>